refactoring app/play/play.tsx
